refactor: 메뉴 및 판매 관리 컴포넌트 FSD 구조통합, 불필요한 파일 삭제 및 새로운 위젯 추가
