+++ src/app/sales-dashboard/page.tsx
@@ -1,9 +1,9 @@
-import { MonthlyOrderList } from "@/components/pages/sales-dashboard/MonthlyOrderList";
+import SalesManage from "@/views/SalesManage";
 
 export default function DashboardPage() {
   return (
     <main className="flex h-screen items-center p-24 bg-[#F2F2F0]">
-      <MonthlyOrderList />
+      <SalesManage />
     </main>
   );
 }
