--- src/lib/UserResponseType.ts
@@ -1,8 +1,0 @@
-export type UserResponseType = {
-  id: string;
-  email: string;
-  store_name: string;
-  created_at: string;
-  updated_at: string;
-  phone: string;
-};
