+++ src/widgets/index.ts
@@ -1,4 +1,6 @@
+export { Calendar } from "./calendar/ui/Calendar";
 export { ConfirmedOrderList } from "./confirmed-order-list/ui/ConfirmedOrderList";
 export { InprogressOrderList } from "./inprogress-order-list/ui/InprogressOrderList";
 export { MenuCart } from "./menu-cart/ui/MenuCart";
 export { MenuList } from "./menu-list/ui/MenuList";
+export { OrderPieChart } from "./order-piechart/ui/OrderPieChart";
